Turning detail up to 11
Because browsers are faster these days, I've made so the detail slider
can go up to 11 for more detailed penrose drawings. It will take a while
to render at that level of detail, so be patient. My M1 Macbook Pro will
take a good 20 seconds before the page is rendered.

I've also upgraded a couple of dependencies to make it work under
python3 and with a never version of underscore. Hurray.